backend/requirements.txt
asgiref==3.11.0
boto3==1.34.0
botocore==1.34.162
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==10.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
s3transfer==0.9.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
